🔵 mv and rename splitServerEndpoint
