+++ cmd/test/api_mapping_test.go
@@ -82,7 +82,7 @@ func TestSplitServerEndpoint(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			guessedServer, guessedEndpoint, err := SplitServerEndpoint(tc.method, tc.requestURL)
+			guessedServer, guessedEndpoint, err := endpoint.SplitServerEndpoint(tc.method, tc.requestURL)
 			if tc.expectError != (err != nil) {
 				t.Fail()
 			}
